+++ @types/index.d.ts
@@ -1,22 +1,46 @@
+/**
+ * 
+ */
 export interface CityCity {
   hotelCity: string;
 }
 
-export interface GoogleCloud {
-  agentModel: string;
-  apiEndPoints: any;
-  projectIdentity: string;
-  projectLocation: string;
+/**
+ * 
+ */
+export interface Coordinates {
+  latitude: number | null;
+  longitude: number | null;
+  country_code: string | null;
 }
 
+/**
+ * 
+ */
 export interface InterpretedInfo {
-  city: string;
-  country_code: string;
-  check_in: string;
-  check_out: string;
-  num_of_rooms: string;
+  city?: string;
+  country_code?: string;
+  check_in?: string;
+  check_out?: string;
+  num_of_rooms?: string;
 }
 
+/**
+ * 
+ */
+export interface InterpretedResponse {
+  city?: string;
+  country_code?: string;
+  check_in?: string;
+  check_out?: string;
+  num_of_rooms?: string;
+  latitude?: number;
+  longitude?: number;
+}
+
+/**
+ * 
+ */
 export interface Metadata {
   title: string;
   description: string;
@@ -24,13 +48,48 @@ export interface Metadata {
   authors: any;
 }
 
-export interface SendOptions {
-  instances: any[];
-  parameters?: any;
+/**
+ * 
+ */
+export interface ParsedInput {
+  location?: string;
+  start_date?: string;
+  end_date?: string;
+  num_of_guests?: number;
+  num_of_rooms?: number;
 }
 
-export interface ServiceInputResponse {
-  example?: string; // Example message
-  failure?: string; // Failure message
-  success?: string; // Success message
+/**
+ * 
+ */
+export interface SearchBarProps {
+  handleQuery: (query: string) => void;
 }
+
+/**
+ * 
+ */
+export interface SearchResultProps {
+  searchResult: InterpretedResponse;
+}
+
+/**
+ * 
+ */
+export interface ServiceInputResponse {
+  example?: string;
+  failure?: string;
+  success?: string;
+}
+
+/**
+ * 
+ */
+export interface writeOptions {
+  modelId: string;
+  response: any;
+}
+
+/***
+  @COPYRIGHT (c) 2024. Thomas EC. Smith (https://www.TECSmith.uk). All rights reserved.
+****/