+++ client/src/utils/setServiceType.ts
@@ -9,7 +9,7 @@ import { ServiceInputResponse } from '../../../@types';
 export const setServiceType = (value: { serviceType: string }): ServiceInputResponse => {
   const { serviceType } = value;
 
-  // Validate the input against the defined schema
+  // Validate the input against the defined schema.
   const validatedField = BookingServiceSchema.safeParse({ serviceType });
 
   if (!validatedField.success) {
@@ -18,13 +18,13 @@ export const setServiceType = (value: { serviceType: string }): ServiceInputResp
 
   switch (serviceType) {
     case 'Hotels':
-      return { example: 'Try; Stockholm from 18th to 28th of March.' };
+      return { example: 'Try; Stockholm, Sverige from 18th to 28th of March for one person.' };
     case 'Flights':
-      return { example: 'Try; Stockholm to Copenhagen from 10th to 20th of April.' };
+      return { example: 'Try; Stockholm to Copenhagen from 10th to 20th of April for one person.' };
     case 'Trains':
-      return { example: 'Try; Stockholm to Uppsala from 7th to 17th of May.' };
+      return { example: 'Try; Stockholm to Uppsala from 7th to 17th of May for one person.' };
     case 'Cars':
-      return { example: 'Try; Stockholm from 1st to 11th of June.' };
+      return { example: 'Try; Stockholm, Sverige from 1st to 11th of June for one person.' };
     default:
       throw new Error('Service type must be one of the following: Hotels, Flights, Trains, Cars.');
   }